Put every shared test fixture in the harness
The fixtures a second subject reached for were living beside the first one
that needed them: withPlatform in platform_test.go, and the listing reader in
api_test.go. Both are now in harness_test.go, where a fixture used by more than
one subject belongs.

The database helper that reads an endpoint's stored UUIDs was named uuidsFor,
which is also a socketRegistry method answering a different question about the
same ID. It reads the store, so it is storedUUIDs.

The delete tests now follow api.go's own order, and the asset helper sits above
the tests that use it, as helpers do in every other file here.
